bdmejora

+++ Backend/Backend/app/Http/Controllers/ofertaController.php
@@ -13,9 +13,11 @@ class ofertaController extends Controller
      */
     public function index()
     {
-        $sql = 'SELECT * FROM oferta';
-        $oferta = DB::select($sql);
-        return $oferta;
+       // Obtener todos las ofertas de la base de datos
+       $oferta = oferta::all();
+
+       // Retornar los oferta como respuesta
+       return response()->json(['ofeta' => $oferta], 200);
     }
 
    
@@ -25,6 +27,7 @@ class ofertaController extends Controller
      */
     public function store(Request $request)
     {
+        /*
         $rules = [
             'Producto' => 'required|min:2|max:20',
             'Precio de Venta(bs)' =>'required|numeric|money',
@@ -34,13 +37,15 @@ class ofertaController extends Controller
         ];
     
         $validatedData = $request->validate($rules);
-    
+    */
         $oferta = new oferta();
         $oferta->descripcion = $request->input('Descripción');
         $oferta->fechaIni = $request->input('Inicio de oferta');
         $oferta->fechaFin = $request->input('Fin oferta');
         $oferta->precioventa = $request->input('Precio de Venta(bs)');
         $oferta->save();
+         // Retornar una respuesta de éxito
+         return response()->json(['mensaje' => 'Producto creado con éxito'], 201);
     }
 
     /**
@@ -56,7 +61,30 @@ class ofertaController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        return oferta::find($id);
+        $oferta = oferta::find($id);
+
+    if (!is_null($oferta)) {
+
+     
+
+    // Actualizar los datos del oferta con los datos del formulario
+    $oferta->codprod = $request->input('codprod');
+    $oferta->descripcion = $request->input('descripcion');
+    $oferta->fechaentrada = $request->input('fechaentrada');
+    $oferta->fechavencimiento = $request->input('fechavencimiento');
+    $oferta->precioVenta = $request->input('precioVenta');
+   
+
+    // Guardar los cambios en la base de datos
+    $oferta->save();
+
+    // Retornar una respuesta de éxito
+    return response()->json(['mensaje' => 'oferta actualizado con éxito'], 200);
+    }
+
+    else{
+        return response()->json(['mensaje' => 'Producto no encontrado'], 404);
+    }
     }
 
     /**
@@ -64,7 +92,17 @@ class ofertaController extends Controller
      */
     public function destroy(string $id)
     {
-        $oferta = oferta::find($id);
-        $oferta->delete();
+            // Encuentra la categoría por su ID
+            $oferta = oferta::find($id);
+            // Verifica si la categoría existe
+            if (!$oferta) {
+               return response()->json(['mensaje' => 'oferta no encontrado'], 404);
+           }
+   
+           // Realiza la eliminación
+           $oferta->delete();
+   
+           // Retorna una respuesta
+           return response()->json(['mensaje' => 'oferta eliminado'], 200);
     }
 }
